Invert Y, Splash, Skills
Player can invert Y mouse from settings that is saved throughout the scenes as well as skills explanations that the player can refer back to in the settings menu. Splash screens shows team logo on startup of the game. Had to completely remake the Main Menu from scratch because it broke the first time.